requirements.txt
streamlit
Pillow
requests
json
io
os